+++ packages/expo-video/src/VideoPlayer.types.ts
@@ -171,11 +171,12 @@ export type VideoPlayerStatus = 'idle' | 'loading' | 'readyToPlay' | 'error';
 
 export type VideoSource =
   | string
+  | number
   | {
       /**
-       * The URI of the video.
+       * The URI of the video or a result of a `require` call for local media sources.
        */
-      uri: string;
+      uri: string | number;
       /**
        * Specifies the DRM options which will be used by the player while loading the video.
        */